+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ***Added:***
 
 - Add the ability to externally manage pip with the `PYAPP_PIP_EXTERNAL` and `PYAPP_PIP_VERSION` options
+- Allow for project installation with a dependency file using the  `PYAPP_PROJECT_DEPENDENCY_FILE` option
 
 ## 0.7.0 - 2023-05-24